Add files via upload
Attempting to complete verse 1. Changed back to original beep sound because new technique of reverb increased overall quality of the piece. Using the wide range of notes 'bb4' and learning the use of 'end' and how it is needed for a complete code to run. Attemping to fasten pace to better mimic the song.