Add files via upload
Changed direction of the remix to be more inclusive of creativity and originality rather than just a cover. Incorporating multiple samples to create various layers and doing it gradually to demonstrate the complexity. Turned it into a hardstyle, techno remix without deleting the initial lofi section. It became an intentional creative choice to demonstrate the various sounds created in songs due to developments in technology.  Final tweaks were involved in the refinement process and continued attempting to use and learn about loops and getting these to play together and in order.